some fixes

+++ FeedUpdater/PodcastManager.FeedUpdater.Domain/Models/Enclosure.cs
@@ -1,3 +1,3 @@
 namespace PodcastManager.FeedUpdater.Domain.Models;
 
-public record Enclosure(string Url, int Length, string Type);
+public record Enclosure(string Url, long Length, string Type);
